util/util: generate stable names
We tried to use random names to prevent clashes with existing paths.
This led to a mismatch in the Mount and Umount stages becaue a
different path got generated.

Assuming that the same calls are done each time, just enumerate the
entries to make them stable on rerun for the same configuration file.