Force filesystem format for OEM partition to be btrfs
In Flatcar's Ignition branch there was a function to try all supported
filesystem in case the user passed ext4 for the OEM partition while it
is using btrfs. Since the OEM partition is going to be reused as is,
we should handle the wrong format gracefully instead of failing now
because the upstream Ignition version doesn't have this fallback logic.
Silently enforce that the format of the OEM partition is btrfs so that
even if the user has an outdated config that uses ext4, it works.